OPL-11128 - Update Delete Documentation
Much of the data delete documentation was either out of date or
incorrect.

Removed the NNT delete documentation as that is no longer used and
updated the full delete documentation to be correct.

It is likely there is still some incorrect data in the text and
histogram delete calls, as things have changed a lot without changing
the existing documentation, but this fixes the known issues that I
found.